requirements.txt
numpy
matplotlib.pyplot
matplotlib.animation
pandas
seaborn
scipy.stats
scipy
os
time